+++ src/modules/email-validation/index.ts
@@ -1,0 +1,2 @@
+export { SkyEmailValidationDirective } from './email-validation.directive';
+export { SkyEmailValidationModule } from './email-validation.module';