+++ src/modules/numeric/index.ts
@@ -1,0 +1,3 @@
+export { SkyNumericPipe } from './numeric.pipe';
+export { SkyNumericService } from './numeric.service';
+export { SkyNumericModule } from './numeric.module';